requirements.txt
numpy>=1.24
matplotlib>=3.7
Pillow>=9.5
scikit-image>=0.21
torch>=2.0
torchvision>=0.15
